fix: merge modules inside single file & resume di
